requirements.txt
# LiveKit Agents
livekit-agents
livekit-agents[cartesia,silero,openai]

# FastAPI and Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8

# AI and ML Libraries
numpy==1.24.3

# Vector Database and Embeddings
sentence-transformers==2.7.0

# AWS Services
boto3==1.34.0
botocore==1.34.0

# File Processing and Resume Parsing
python-magic==0.4.27
pypdf==3.17.1

# HTTP and API Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and Configuration
python-dotenv==1.0.0

# CORS and Middleware
fastapi-cors==0.0.6

# Time and Date
python-dateutil==2.8.2
pytz==2023.3

# JSON and Data Processing
orjson==3.9.10

# Async Support
aiofiles==23.2.1

# Database Connection Pooling
asyncpg==0.29.0

# Rate Limiting
slowapi==0.1.9

# Health Checks
fastapi-health==0.1.0
